Problem: transaction validation code is split in two parts (CRO-175)
Solution: WIP created a crate and moved the chain core's witness verification into it
