+++ Cargo.toml
@@ -3,6 +3,7 @@
 members = [
     "chain-abci",
     "chain-core",
+    "chain-tx-validation",
     "client-cli",
     "client-common",
     "client-core",
