+++ chain-abci/Cargo.toml
@@ -9,6 +9,7 @@ edition = "2018"
 [dependencies]
 abci = "0.5.4"
 chain-core = { path = "../chain-core" }
+chain-tx-validation = { path = "../chain-tx-validation" }
 log = "0.4.0"
 env_logger = "0.6.1"
 bit-vec = "0.6.1"
