+++ chain-tx-validation/src/lib.rs
@@ -1,0 +1,2 @@
+/// transaction witness verification
+pub mod witness;
